Update cache.clj
The problem with a simple timeout is that the 'progress complete' report can be printed after other work has continued. This is confusing for users and is a result of a naive timeout in a loop. Instead, use an async channel to communicate with progress printing. This means we don't need to print with such high frequency, and we can immediately print when completed while waiting for a timeout to update the progress display.